+++ src/database.rs
@@ -121,7 +121,7 @@ impl Database {
                 && name_str == name {
                 true
             }
-            else if let Id::CsType(cs_name) = &a.id && cs_name.ends_with(name) {
+            else if let Id::CsType { name: n, .. } = &a.id && n == name {
                 return true;
             }
             else {
